backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
llama-cpp-python
requests
PyPDF2
